requirements.txt
anytree
numpy
OneEuroFilter
opencv-python
onnxruntime-gpu
torch
tqdm